Implement RegenMeter into StatusOrbs
Also removes the RegenMeter plugin
